Make calendar event type byte instead of int
Now the UUID will be constructed like this:

High 64bit 0x4742474200 | type
Low 64bit id

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/EventHandler.java
@@ -55,5 +55,5 @@ public interface EventHandler {
 
     void onAddCalendarEvent(CalendarEventSpec calendarEventSpec);
 
-    void onDeleteCalendarEvent(int type, long id);
+    void onDeleteCalendarEvent(byte type, long id);
 }
