+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/model/CalendarEventSpec.java
@@ -1,11 +1,11 @@
 package nodomain.freeyourgadget.gadgetbridge.model;
 
 public class CalendarEventSpec {
-    public static final int TYPE_UNKNOWN = 0;
-    public static final int TYPE_SUNRISE = 1;
-    public static final int TYPE_SUNSET = 2;
+    public static final byte TYPE_UNKNOWN = 0;
+    public static final byte TYPE_SUNRISE = 1;
+    public static final byte TYPE_SUNSET = 2;
 
-    public int type;
+    public byte type;
     public long id;
     public int timestamp;
     public int durationInSeconds;
